Extract CloudFormationFixture helper methods to new class
